rename chatItems to chatMessage folder
