+++ apps/src/aiComponentLibrary/chatMessage/chat-message.module.scss
