+++ apps/src/aichat/redux/aichatRedux.ts
@@ -7,7 +7,7 @@ import {
   ThunkDispatch,
 } from '@reduxjs/toolkit';
 
-import {Role} from '@cdo/apps/aiComponentLibrary/chatItems/types';
+import {Role} from '@cdo/apps/aiComponentLibrary/chatMessage/types';
 import Lab2Registry from '@cdo/apps/lab2/Lab2Registry';
 import {PLATFORMS} from '@cdo/apps/lib/util/AnalyticsConstants';
 import analyticsReporter from '@cdo/apps/lib/util/AnalyticsReporter';
